fix(bl31): typecast operands to match data type
This corrects the MISRA violation C2012-10.3:
The value of an expression shall be assigned with an essential type
category.

State: not-upstreamable (fusa)
Link: https://review.trustedfirmware.org/c/TF-A/trusted-firmware-a/+/40540

+++ include/plat/common/platform.h
@@ -139,7 +139,7 @@ uint32_t plat_ic_acknowledge_interrupt(void);
 uint32_t plat_ic_get_interrupt_type(uint32_t id);
 void plat_ic_end_of_interrupt(uint32_t id);
 uint32_t plat_interrupt_type_to_line(uint32_t type,
-				     uint32_t security_state);
+				     size_t security_state);
 
 /*******************************************************************************
  * Optional interrupt management functions, depending on chosen EL3 components.
